api_yamdb/requirements.txt
attrs==22.1.0
bcrypt==4.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.3
distro==1.8.0
Django==3.2.16
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.2.0
docker==6.0.1
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
flake8==5.0.4
gunicorn==20.0.4        
idna==3.4
iniconfig==1.1.1        
jsonschema==3.2.0       
mccabe==0.7.0
packaging==21.3
paramiko==2.12.0        
pluggy==0.13.1 
py==1.11.0
pycodestyle==2.9.1      
pycparser==2.21
pyflakes==2.5.0
PyJWT==2.1.0
PyNaCl==1.5.0
pyparsing==3.0.9        
pyrsistent==0.19.2
psycopg2-binary==2.9.5 
pytest==6.2.4
pytest-django==4.4.0    
pytest-pythonpath==0.7.3
python-dotenv==0.21.0
python-dotenv==0.21.0   
pytz==2022.2.1
PyYAML==5.4.1
requests==2.28.1        
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4        
toml==0.10.2
urllib3==1.26.12        
websocket-client==0.59.0